Completed layouts of home page

+++ src/data/schemas/schema-0.ts
@@ -9,7 +9,7 @@ export class Expense {
     properties: {
       id: 'string',
       type: 'string',
-      amount: 'int',
+      amount: 'float',
       date: 'date'
     }
   }
